Model fields for automatically updated releases
The model fields `discogs_release_id` and `last_checked_at` on a Release
will be used to track the id of the corresponding Discogs release and
when this Discogs release was last checked for updates.

They will be used in the following management command for automatically
updating Releases based on changes in the corresponding Discogs release.